Fix TestPipeline passing `no-content` to Adapter when message was empty (#11204)
